pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-magnetgeo"
version = "1.0.0"
description = "Python helpers to create HiFiMagnet cads and meshes"
readme = "README.rst"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Christophe Trophime", email = "christophe.trophime@lncmi.cnrs.fr"},
    {name = "Romain Vallet", email = "romain.vallet@lncmi.cnrs.fr"},
    {name = "Jeremie Muzet", email = "jeremie.muzet@lncmi.cnrs.fr"}
]
maintainers = [
    {name = "Christophe Trophime", email = "christophe.trophime@lncmi.cnrs.fr"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]
keywords = ["magnet", "geometry", "cad", "mesh", "simulation", "hifimagnet"]
dependencies = [
    "pyyaml>=6.0.2,<7.0.0",
    "pandas>=1.5.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "tox>=4.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.22.0",
]
test = [
    "pytest>=8.2.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Trophime/python_magnetgeo"
Documentation = "https://python-magnetgeo.readthedocs.io"
Repository = "https://github.com/Trophime/python_magnetgeo"
Issues = "https://github.com/Trophime/python_magnetgeo/issues"
Changelog = "https://github.com/Trophime/python_magnetgeo/blob/main/HISTORY.rst"

[project.scripts]
load-profile-from-dat = "python_magnetgeo.examples.load_profile_from_dat:main"
split-helix-yaml = "python_magnetgeo.examples.split_helix_yaml:main"
check-magnetgeo-yaml = "python_magnetgeo.examples.check_magnetgeo_yaml:main"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["python_magnetgeo*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
python_magnetgeo = ["py.typed"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=python_magnetgeo",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests for individual classes",
    "integration: Integration tests across multiple classes",
    "serialization: Tests for JSON/YAML serialization",
    "geometric: Tests for geometric operations",
    "probe: Tests for probe system functionality",
    "slow: Tests that take longer to run",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["python_magnetgeo"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Black code formatter
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort import sorter
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# MyPy type checker
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Tox test automation
[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py311,py312,flake8,mypy
isolated_build = true
skip_missing_interpreters = true

[testenv]
deps =
    pytest>=8.2.0
    pytest-cov>=4.0.0
commands =
    pytest {posargs}

[testenv:flake8]
deps = flake8>=6.0.0
commands = flake8 python_magnetgeo tests

[testenv:mypy]
deps =
    mypy>=1.0.0
    types-PyYAML
commands = mypy python_magnetgeo

[testenv:black]
deps = black>=23.0.0
commands = black --check python_magnetgeo tests

[testenv:docs]
deps =
    sphinx>=6.0.0
    sphinx-rtd-theme>=1.2.0
changedir = docs
commands = sphinx-build -W -b html -d {envtmpdir}/doctrees . {envtmpdir}/html
"""

# Ruff linter (alternative to flake8)
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["F401", "F811"]

[tool.ruff.isort]
known-first-party = ["python_magnetgeo"]
